Add XML docs
